+++ src/atoms/error.ts
@@ -1,10 +1,15 @@
 import { ReactNode } from "react";
-import errorAtom from "./atom";
+import { atom } from "recoil";
 
 export type ErrorType = {
   title: string;
   message: ReactNode;
   record: Nullable<string>;
 };
+
+const errorAtom = atom<Nullable<ErrorType>>({
+  key: "errorAtom",
+  default: null,
+});
 
 export default errorAtom;
