+++ src/atoms/isVirtualKeyboardDetected.ts
