+++ src/components/AddRoom/AddRoom.tsx
@@ -15,8 +15,8 @@ import OptionDate from "components/common/roomOptions/Date";
 import OptionTime from "components/common/roomOptions/Time";
 import OptionMaxPeople from "components/common/roomOptions/MaxPeople";
 
-import alertAtom from "recoil/alert";
-import myRoomAtom from "recoil/myRoom";
+import alertAtom from "atoms/alert";
+import myRoomAtom from "atoms/myRoom";
 
 import { date2str, getToday10, getToday } from "tools/moment";
 import theme from "styles/theme";
