+++ src/components/Chatting/Header/Popup/PopupPay.jsx
@@ -2,8 +2,8 @@ import { useAxios } from "hooks/useTaxiAPI";
 import PopupContainer from "./PopupContainer";
 import PropTypes from "prop-types";
 import { useSetRecoilState } from "recoil";
-import alertAtom from "recoil/alert";
-import myRoomAtom from "recoil/myRoom";
+import alertAtom from "atoms/alert";
+import myRoomAtom from "atoms/myRoom";
 
 const PopupPay = (props) => {
   const axios = useAxios();
