+++ src/components/Home/InfoSection.tsx
@@ -2,8 +2,8 @@ import { useMemo } from "react";
 import { Link } from "react-router-dom";
 import RLayout from "components/common/RLayout";
 import { useRecoilValue } from "recoil";
-import loginInfoDetailAtom from "recoil/loginInfoDetail";
-import myRoomAtom from "recoil/myRoom";
+import loginInfoDetailAtom from "atoms/loginInfoDetail";
+import myRoomAtom from "atoms/myRoom";
 import theme from "styles/theme";
 import moment, { getToday } from "tools/moment";
 import Button from "components/common/Button";
