+++ src/components/Search/RoomSelectionModal.jsx
@@ -2,9 +2,9 @@ import { useEffect, useState, useRef } from "react";
 import { useHistory } from "react-router-dom";
 import { useAxios } from "hooks/useTaxiAPI";
 import { useRecoilValue, useSetRecoilState, useRecoilState } from "recoil";
-import loginInfoDetailAtom from "recoil/loginInfoDetail";
-import alertAtom from "recoil/alert";
-import myRoomAtom from "recoil/myRoom";
+import loginInfoDetailAtom from "atoms/loginInfoDetail";
+import alertAtom from "atoms/alert";
+import myRoomAtom from "atoms/myRoom";
 import PropTypes from "prop-types";
 import { date2str } from "tools/moment";
 import { getLocationName } from "tools/trans";
