+++ src/components/Skeleton/ChannelTalkProvider/index.tsx
@@ -1,6 +1,6 @@
 import { useEffect } from "react";
 import { useRecoilValue } from "recoil";
-import loginInfoDetailAtom from "recoil/loginInfoDetail";
+import loginInfoDetailAtom from "atoms/loginInfoDetail";
 import ChannelService from "./channelService";
 import { channelTalkPluginKey } from "serverconf";
 import { useLocation } from "react-router-dom";
