+++ src/components/Skeleton/Navigation.tsx
@@ -5,7 +5,7 @@ import hoverEventSet from "tools/hoverEventSet";
 import theme from "styles/theme";
 
 import { useRecoilValue } from "recoil";
-import isVirtualKeyboardDetectedAtom from "recoil/isVirtualKeyboardDetectedAtom";
+import isVirtualKeyboardDetectedAtom from "atoms/isVirtualKeyboardDetected";
 
 import RoofingRoundedIcon from "@mui/icons-material/RoofingRounded";
 import SearchRoundedIcon from "@material-ui/icons/SearchRounded";
