+++ src/hooks/useTaxiAPI/useAxios.tsx
@@ -2,7 +2,7 @@ import { useCallback } from "react";
 import { useHistory, useLocation } from "react-router-dom";
 import { useSetRecoilState } from "recoil";
 
-import errorAtom from "recoil/error";
+import errorAtom from "atoms/error";
 import axios from "./axios";
 
 type AxiosOption = {
