--- src/recoil/alert/index.ts
@@ -1,3 +1,0 @@
-import alertAtom from "./atom";
-
-export default alertAtom;
