--- src/recoil/loginInfoDetail/atom.ts
@@ -1,9 +1,0 @@
-import { atom } from "recoil";
-import { LoginInfoDetailType } from "./index";
-
-const loginInfoDetailAtom = atom<Nullable<LoginInfoDetailType>>({
-  key: "loginInfoDetailAtom",
-  default: null,
-});
-
-export default loginInfoDetailAtom;
