--- src/recoil/myRoom/atom.ts
@@ -1,9 +1,0 @@
-import { atom, RecoilState } from "recoil";
-import { MyRoomType } from "./index";
-
-const myRoomAtom = atom<Nullable<MyRoomType>>({
-  key: "myRoomAtom",
-  default: null,
-});
-
-export default myRoomAtom;
