--- src/recoil/taxiLocation/index.js
@@ -1,3 +1,0 @@
-import taxiLocationAtom from "./atom";
-
-export default taxiLocationAtom;
